# backends/internal/transport/resthttp/swaggerui/swagger.yaml
basePath: /api
consumes:
- application/json
definitions:
...
  Role:
    format: uint64
    type: integer
    x-go-package: github.com/rasulov-emirlan/todo-app/backends/internal/domain/users
...
  reqUsersRefresh:
    description: reqUsersRefresh is used for mobile clients. They should send their
      refresh keys in this model to refresh endpoint for updating their keys
    properties:
      refreshKey:
        type: string
        x-go-name: RefreshKey
    type: object
    x-go-package: github.com/rasulov-emirlan/todo-app/backends/internal/transport/resthttp
  reqUsersSignIn:
    description: |-
      reqUsersSignIn
      represents all info needed for a user to sign in
    properties:
      email:
        description: the email address for this user
        example: user@example.com
        type: string
        x-go-name: Email
      password:
        description: password for this user
        example: password
        type: string
        x-go-name: Password
    required:
    - email
    - password
    type: object
    x-go-package: github.com/rasulov-emirlan/todo-app/backends/internal/transport/resthttp
  reqUsersSignUp:
    description: |-
      reqUsersSignUp
      represents all info needed for a user to sign up
      It is used only for signing up
    properties:
      email:
        description: the email address for this user
        example: user@example.com
        type: string
        x-go-name: Email
      password:
        description: the password for this user
        example: password
        maxLength: 128
        minLength: 6
        type: string
        x-go-name: Password
      username:
        description: the name for this user
        example: John Doe
        maxLength: 20
        minLength: 6
        type: string
        x-go-name: Username
    required:
    - email
    - password
    - username
    type: object
    x-go-package: github.com/rasulov-emirlan/todo-app/backends/internal/transport/resthttp
...
  stdResponse:
    description: |-
      stdResponse
      represents the wrapper that all responses come inside of.
    properties:
      data:
        description: Actual Data that you expect to get on success. It will be omited
          if any errors occure
        x-go-name: Data
      errors:
        description: Error messages. It will not be omited if null
        items:
          type: string
        type: array
        x-go-name: Errors
    required:
    - errors
    - data
    type: object
    x-go-package: github.com/rasulov-emirlan/todo-app/backends/internal/transport/resthttp
...
  usersMeResponse:
    description: user represents a response for /users/me endpoint
    properties:
      createdAt:
        format: date-time
        type: string
        x-go-name: CreatedAt
      email:
        type: string
        x-go-name: Email
      id:
        type: string
        x-go-name: ID
      role:
        $ref: '#/definitions/Role'
      updatedAt:
        format: date-time
        type: string
        x-go-name: UpdatedAt
      username:
        type: string
        x-go-name: Username
    type: object
    x-go-package: github.com/rasulov-emirlan/todo-app/backends/internal/transport/resthttp
info:
  contact:
    email: rasulov.emirlan@gmail.com
    name: Emirlan Rasulov
    url: https://github.com/rasulov-emirlan
  description: |-
    the purpose of this application is to learn more about REST and swagger

    This should demonstrate how to write clean code in go
    and communicate with it using http
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
  termsOfService: there are no TOS at this moment, use at your own risk we take no
    responsibility
  title: Todo App API
  version: 0.0.1
paths:
...
  /users/auth/refresh:
    post:
      consumes:
      - application/json
      description: |-
        This is supposed to return a new pair of keys. It will check the body for refresh key.
        If it wont find it in body, then it will check refresh_key Cookie. If both are empty then
        ur mad bro.
      operationId: UsersRefresh
      parameters:
      - description: Refresh key
        in: body
        name: refresh
        schema:
          $ref: '#/definitions/reqUsersRefresh'
          description: Refresh key
          type: object
      - description: Refresh Key
        name: refreshCookie
        type: string
      produces:
      - application/json
      responses:
        "200":
          $ref: '#/responses/usersKeys'
        "422":
          description: stdResponse
          schema:
            $ref: '#/definitions/stdResponse'
        default:
          $ref: '#/responses/usersKeys'
      schemes:
      - http
      - https
      summary: Refresh keys
      tags:
      - auth
  /users/auth/signin:
    post:
      consumes:
      - application/json
      description: This should return a pair of keys for the user, if user info provided
        is valid
      operationId: UsersSignIn
      parameters:
      - description: Basic info for user to sign up
        in: body
        name: user info
        required: true
        schema:
          $ref: '#/definitions/reqUsersSignIn'
          description: Basic info for user to sign up
          type: object
      produces:
      - application/json
      responses:
        "200":
          $ref: '#/responses/usersKeys'
        "422":
          description: stdResponse
          schema:
            $ref: '#/definitions/stdResponse'
        default:
          $ref: '#/responses/usersKeys'
      schemes:
      - http
      - https
      summary: Sign in a user
      tags:
      - auth
  /users/auth/signup:
    post:
      consumes:
      - application/json
      description: |-
        This will create a user in our database IF AND ONLY
        if he doesnt exist yet. After creating him it will automaticaly
        sign him in
      operationId: UsersSignUp
      parameters:
      - description: Basic info for user to sign up
        in: body
        name: user info
        required: true
        schema:
          $ref: '#/definitions/reqUsersSignUp'
          description: Basic info for user to sign up
          type: object
      produces:
      - application/json
      responses:
        "200":
          $ref: '#/responses/usersKeys'
        "422":
          description: stdResponse
          schema:
            $ref: '#/definitions/stdResponse'
        default:
          $ref: '#/responses/usersKeys'
      schemes:
      - http
      - https
      summary: Sign up a user
      tags:
      - auth
  /users/me:
    get:
      consumes:
      - application/json
      description: This will return current user.
      operationId: UsersMe
      produces:
      - application/json
      responses:
        default:
          description: usersMeResponse
          schema:
            $ref: '#/definitions/usersMeResponse'
      schemes:
      - http
      - https
      security:
      - Bearer:
        - '[]'
      summary: Get current user
      tags:
      - users
  /users{id}:
    delete:
      consumes:
      - application/json
      description: This will delete delete user.
      operationId: UsersDelete
      parameters:
      - description: Id of the user you wish to delete
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        default:
          description: stdResponse
          schema:
            $ref: '#/definitions/stdResponse'
      schemes:
      - http
      - https
      security:
      - Bearer:
        - '[]'
      summary: Delete a user
      tags:
      - users
produces:
- application/json
- application/xml
responses:
  usersKeys:
    description: usersKeys represents a pair of keys used for authorization in our
      api
    headers:
      refresh_key:
        description: |-
          This cookie will contain refresh key

          in: cookie
        type: string
schemes:
- http
- https
security:
- Bearer:
  - '[]'
securityDefinitions:
  Bearer:
    in: Header
    name: Authorization
    type: apiKey
swagger: "2.0"